// clients/accountingService/models/CostCentreBudgetCreateDto.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type CostCentreBudgetCreateDto = {
    readonly id?: string;
    readonly timestamp?: string;
    name?: string | null;
    tenantId?: string | null;
    fiscalYearId?: string | null;
    costCentreId?: string | null;
};