// clients/accountingService/models/CostCentreGroupCreateDto.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type CostCentreGroupCreateDto = {
    readonly id?: string;
    readonly timestamp?: string;
    name?: string | null;
    description?: string | null;
    disabled?: boolean;
    tenantId?: string | null;
    parentCostCentresGroupId?: string | null;
};