// clients/accountingService/models/CostCentreGroupUpdateDto.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type CostCentreGroupUpdateDto = {
    name?: string | null;
    description?: string | null;
    disabled?: boolean;
    tenantId?: string | null;
    parentCostCentresGroupId?: string | null;
};